Download zip file into ORACLE_ZIP_DIR if present
If ORACLE_ZIP_DIR is set, the zip file is stored in that
directory.

This allows the zip file to be located outside of the source tree,
and thus in a cachable directory as required for the Travis caching
system, which became available on trusty images on 03-05-2016.
https://blog.travis-ci.com/2016-05-03-caches-are-coming-to-everyone

While the Oracle XE download time is negligible, the downloaded
Oracle XE zip is made available to pull requests, whereas currently
download.js fails because ORACLE_LOGIN_* environment variables are
not available to pull requests.
https://docs.travis-ci.com/user/caching/